Reorganize script structure
