--- Assets/Scripts/CollisionRefrain.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: dc456a365b8354d3ea205b6bc443ae79
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
