--- Assets/Scripts/HopsitalAudioFinal.cs
@@ -1,15 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class HopsitalAudioFinal : MonoBehaviour
-{
-    public AudioSource source;
-    public AudioClip congrats;
-
-
-    void Start()
-    {
-        source.PlayOneShot(congrats);
-    }
-}
