+++ Assets/Scripts/Hospital.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e576f3ee1cb162945b9fa5cd8a4f2a1b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
