+++ Assets/Scripts/Hospital/HospitalManager.cs
