--- Assets/Scripts/PlayerInHospital.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 969aa15a778f940c9accc3786b26f7cc
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
